+++ release/components/input/input.module.js
@@ -6,6 +6,8 @@ import { InputHintDirective } from './input-hint.directive';
 import { AutosizeDirective } from './input-autosize.directive';
 import { InputPrefixComponent } from './input-prefix.component';
 import { InputSuffixComponent } from './input-suffix.component';
+import { MinValidatorDirective } from './validators/min-value.directive';
+import { MaxValidatorDirective } from './validators/max-value.directive';
 var InputModule = (function () {
     function InputModule() {
     }
@@ -16,7 +18,9 @@ var InputModule = (function () {
                         InputHintDirective,
                         AutosizeDirective,
                         InputPrefixComponent,
-                        InputSuffixComponent
+                        InputSuffixComponent,
+                        MinValidatorDirective,
+                        MaxValidatorDirective
                     ],
                     exports: [
                         InputComponent,