+++ release/components/input/validators/max-value.directive.d.ts
@@ -1,0 +1,6 @@
+import { Validator, AbstractControl, ValidationErrors } from '@angular/forms';
+export declare class MaxValidatorDirective implements Validator {
+    max: number;
+    type: string;
+    validate(c: AbstractControl): ValidationErrors | null;
+}