+++ release/components/input/validators/min-value.directive.d.ts
@@ -1,0 +1,6 @@
+import { Validator, AbstractControl, ValidationErrors } from '@angular/forms';
+export declare class MinValidatorDirective implements Validator {
+    min: number;
+    type: string;
+    validate(c: AbstractControl): ValidationErrors | null;
+}